+++ task_management/src/main/java/com/example/service/TimersSettingService.java
@@ -13,8 +13,6 @@ import com.example.entity.TimersSetting;
 import com.example.entity.User;
 import com.example.repository.TimersSettingRepository;
 
-import jakarta.validation.constraints.NotNull;
-
 /**
  * タイマー設定用サービスクラス
  * @author yoshi
@@ -39,7 +37,7 @@ public class TimersSettingService {
 	 * @param 
 	 * 
 	 */
-	public TimersSetting findById(Integer settingId) {
+	public TimersSetting findById(Long settingId) {
 		Optional<TimersSetting> timsersSettingOptional = timersSettingRepository.findById(settingId);
 		return timsersSettingOptional.orElse(null); // OptinalからTimersSettingオブジェクトを取得する
 	}
@@ -71,7 +69,7 @@ public class TimersSettingService {
 	 */
 	public TimersSetting getUsersFocusTimer() {
 		// ログインしているユーザー情報のIDを取得する
-		Integer userId = userService.getCurrentUserId();
+		Long userId = userService.getCurrentUserId();
 		
 		return timersSettingRepository.findTopByUserIdOrderByIdDesc(userId);
 		
@@ -83,7 +81,7 @@ public class TimersSettingService {
 	 * @param Id ユーザーID
 	 * @return ユーザー別登録タイマー情報リスト
 	 */
-	public List<TimersSetting> getTimersByUserId(Integer userId) {
+	public List<TimersSetting> getTimersByUserId(Long userId) {
 		return timersSettingRepository.findByUserId(userId);
 	}
 
@@ -93,7 +91,7 @@ public class TimersSettingService {
 	 * @param id タイマーID
 	 * @return タイマー情報
 	 */
-	public Optional<TimersSetting> get(Integer id) {
+	public Optional<TimersSetting> get(Long id) {
 		return timersSettingRepository.findById(id);
 	}
 
@@ -102,7 +100,7 @@ public class TimersSettingService {
 	 * 
 	 * @param id タイマーID
 	 */
-	public void delete(Integer id) {
+	public void delete(Long id) {
 		timersSettingRepository.deleteById(id);
 	}
 }
